Managed the OS specifics in OS.h Tidying up.. Messing with the LevelEditor.. Just Tidying up mostly..
Removed all Xcode Warnings so that's great :)

+++ trunk/Meatball/level.cpp
@@ -312,7 +312,7 @@ void cLevel :: Parse( char* command, int line )
 	
 	while(*str)
 	{
-		int len = strcspn( str," " );
+		size_t len = strcspn( str," " );
 		parts[i] = (char*)malloc( len + 1 );
 		memcpy( parts[i], str, len );
 		parts[i][len] = 0;
